fix member model
