+++ models/member.js
@@ -1,0 +1,30 @@
+'use strict';
+const { Model } = require('sequelize');
+module.exports = (sequelize, DataTypes) => {
+  class Member extends Model {
+    /**
+     * Helper method for defining associations.
+     * This method is not a part of Sequelize lifecycle.
+     * The `models/index` file will call this method automatically.
+     */
+    static associate(models) {
+      // define association here
+    }
+  }
+  Member.init(
+    {
+      name: DataTypes.STRING,
+      facebookUrl: DataTypes.STRING,
+      instagramUrl: DataTypes.STRING,
+      linkedinUrl: DataTypes.STRING,
+      image: DataTypes.STRING,
+      deletedAt: DataTypes.DATE,
+    },
+    {
+      sequelize,
+      modelName: 'Member',
+      paranoid: true,
+    }
+  );
+  return Member;
+};
